requirements.txt
google-genai>=0.8.0
google-adk>=0.3.0
google-cloud-secret-manager>=2.20.0
google-cloud-firestore>=2.16.0
google-cloud-monitoring>=2.21.0
google-cloud-storage>=2.18.0
opentelemetry-sdk>=1.25.0
opentelemetry-api>=1.25.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-instrumentation-fastapi>=0.46b0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.48.0
pillow>=11.0.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
websockets>=13.0
python-multipart>=0.0.12
locust>=2.31.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
